# gfx1010 optimizations