Reading @IMbackK 's PR #8082 I've noticed that RDNA1 cards can also benefit from a small performance gain just by adjusting the same values as that PR.

| model                          |       size |     params | backend    | ngl |          test |              t/s master| t/s PR | Speedup |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | ---------------: | ---------------: | ------------:
| llama 8B Q5_K - Small          |   5.21 GiB |     8.03 B | ROCm       |  99 |         pp512 |    163.86 ± 0.18 |    276.60 ± 0.41 | 1.69 |

This is still far from the performance pre #7716 (RDNA1 cards suffered a 50% performance drop with that) but it's still a good performance improvement.

Thanks again to @IMbackK for his PR as I wouldn't have noticed this without it.

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High