@JohannesGaessler 2024-07-03
>I've only managed to get this far as setting the return of get_mmq_y_device to 32 boosts the performance to 800 t/s but completely breaks the inference by outputting only gibberish.

`get_mmq_y_host` and `get_mmq_y_device` need to return consistent values. If you only reduce `get_mmq_y_device` the only part of the matrix multiplication will be done which of course leads to a much faster (but wrong) kernel. As of right now both functions are situated next to each other in `mmq.cuh`. If you want you can try tuning the values again, otherwise I would just merge this PR as-is.