@sorasoras 2024-07-03
Does this pr affect RDNA3? I really can use some optimizations.