@JohannesGaessler 2024-07-03
>I still think this isn't the best way to do it because if different values are needed for different cards this can result in something messy, maybe using a normal if statement so that the row isn't too long, however I haven't been able to make it work that way.

Yes, `constexpr` functions do not allow conditional statements in the C++ 11 standard.

>Apparently all the issues I've been having are caused by the check on int8_mma_available not working as intended.

It is working as intended, you are just not using it as intended. The lowercase `constexpr` functions are intended to be used in host code to determine feature availability at runtime. The uppercase preprocessor macros are intended to be used in device code to determine feature availability at compile time. But if you have a separate check for AMD you don't need `INT8_MMA_AVAILABLE` anyways. On NVIDIA `INT8_MMA_AVAILABLE` is effectively `__CUDA_ARCH__ >= CC_TURING` and therefore always true if `__CUDA_ARCH__ >= CC_VOLTA`.