@daniandtheweb 2024-07-03
Okay, I actually wanted to specify for RDNA1 because I wasn't sure of the effects it could have on the pr #8082.