@JohannesGaessler 2024-07-03
Sorry, actually it has to be done the way you had it with an RDNA1 check. On AMD you cannot do a simple check against a number because there is no sensible value for `__CUDA_ARCH__`.